Correctly append list of command results
In the _execute_methods function, if the results are returned as a list
they were being appended to the results list.  So there would be a list
inside the results list, which would later cause an error while trying
to decode the results.  This happens when parsing state against an
Arista device.  See
https://github.com/napalm-automation/napalm-yang/issues/103