Update base._to_dict_leaf function to display lists
In cases where the final leaf element is a list the to dict function was
not properly translating the lists.

The trunk vlans in the interface model and the ip addresses under the
systems host entries are two examples.

This change will set the value to the `_list` element if it is present
otherwise continue on with current behavior